Blocking GetExecBatchResults (#187)
* Blocking GetExecBatchResults

* Handle optional timeouts

* Use a notifier instead of a delay_for - backed loop

* Replace the loop with Notify::when(..).await

* Remove crossbeam-channel dependency
Remove redundant error types

* Update GSB example

* Simplify Notify::notify impl

* Improve readability

* Set initial Notify uid to max usize value

* Client and spec update

* Change remaining timeout values to f32

* [exe-unit] gsb example facilitation: built in GSB + default values for params

* [exe-unit] cmd result: capture stdout and stderr

* [exe-unit] cmd result is not optional

* [exe-unit] env logger setup

* [exe-unit] cmd error to have both: stdout and stderr + fixed sample commands.json

* [exe-unit] example commands.json: file paths must be absolute

* [exe-unit2gsb] CLI options to test timeout + logs readability

* [client] get exec batch result empty on timeout; ya-req:timeout+logs

* [exe-unit] revert unintentional local-exeunits-descriptor.json change

+++ interfaces/client/src/activity/provider.rs
@@ -41,7 +41,7 @@ impl ActivityProviderApi {
     pub async fn get_activity_events(
         &self,
         #[allow(non_snake_case)]
-        timeout: Option<i32>,
+        timeout: Option<f32>,
         #[allow(non_snake_case)]
         maxEvents: Option<i32>,
     ) -> Result<Vec<ProviderEvent>> {
